// src/models/ShoppingModeItem.ts
export interface ShoppingModeItem {
    documentId: string;
    storeId: string;
    userId : string;
    addedToCart: string;
    materialId : string;
    quantity: number;
    pickedQuantity: number;
    productUPC: string;
    productName: string;
    productNameSearch : string;
    productImg: string;
    productRegularPrice: string;
    productSalesPrice: string;
    status: string;
}